convert BatteryMonitor to return Int rather than Float

+++ app/src/main/java/com/egecius/demo_androidplatform/battery/BatteryMonitor.kt
@@ -5,5 +5,5 @@ import kotlinx.coroutines.flow.Flow
 
 interface BatteryMonitor {
 
-    fun getPercentageFlow(): Flow<Float>
+    fun getPercentageFlow(): Flow<Int>
 }
